Fixed tests

+++ src/main/java/app/config/FileConfigurationFile.java
@@ -2,11 +2,11 @@ package app.config;
 
 import java.io.*;
 
-public class WorkingDirectoryConfigurationFile implements ConfigurationFile {
+public class FileConfigurationFile implements ConfigurationFile {
     private final File file;
 
-    public WorkingDirectoryConfigurationFile(String filename) {
-        this.file = new File(filename);
+    public FileConfigurationFile(File file) {
+        this.file = file;
     }
 
     @Override
@@ -26,8 +26,12 @@ public class WorkingDirectoryConfigurationFile implements ConfigurationFile {
     @Override
     public OutputStream getOutputStream() {
         try {
+            boolean alreadyExists = file.createNewFile();
+            if (!alreadyExists) {
+                throw new RuntimeException("File already exists");
+            }
             return new FileOutputStream(file);
-        } catch (FileNotFoundException e) {
+        } catch (IOException e) {
             throw new ConfigurationFileException(e);
         }
     }
